refactored performance metrics code
