+++ setup.py
@@ -12,7 +12,7 @@ config = {
     'entry_points': {'console_scripts': ['kerasAC_train = kerasAC.train:main',
                                          'kerasAC_predict=kerasAC.predict:main',
                                          'kerasAC_curves=kerasAC.curves:main',
-                                         'kerasAC_score=kerasAC.performance_metrics:main',
+                                         'kerasAC_score=kerasAC.performance_metrics.performance_metrics:main',
                                          'kerasAC_interpret=kerasAC.interpret:main',
                                          'kerasAC_plot_interpretation=kerasAC.plot_interpretation:main',
                                          'kerasAC_cross_validate=kerasAC.cross_validate:main']},
